Function aoc_2015_day_15::logic::iter_recipes

pub fn iter_recipes<'dat>(
    input: &'dat Input<'_>
) -> impl Iterator<Item = NumResult<Recipe>> + 'dat

Iterate through all recipes containing exactly 100 ingredients

We iterate through all values for all but the last ingredient, starting with all zero, then increasing the last ingredient which can be increased without exceeding 100 total, and setting all subsequent ingredients to zero. This gives for example, for four ingredients, 0, 0, 0, then 0, 0, 1, etc…

The last ingredient we always set specifically to get exactly 100 ingredients total.